+++ src/main/kotlin/com/dphong/likescash/api/posts/controller/SubmitPostController.kt
@@ -6,7 +6,7 @@ import com.dphong.likescash.api.posts.model.SubmitPostStatus
 import com.dphong.likescash.api.posts.service.SubmitPostService
 import com.dphong.likescash.common.annotation.LoginMember
 import com.dphong.likescash.common.config.auth.MemberFacade
-import com.dphong.likescash.common.reponse.StatusDataResult
+import com.dphong.likescash.common.response.StatusDataResult
 import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.PostMapping
