+++ src/main/kotlin/com/dphong/likescash/api/product/controller/ProductController.kt
@@ -2,7 +2,7 @@ package com.dphong.likescash.api.product.controller
 
 import com.dphong.likescash.api.product.model.ProductDetails
 import com.dphong.likescash.api.product.service.ProductGetter
-import com.dphong.likescash.common.reponse.DataResult
+import com.dphong.likescash.common.response.DataResult
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
