+++ src/test/kotlin/com/dphong/likescash/api/product/controller/ProductControllerTest.kt
@@ -3,7 +3,7 @@ package com.dphong.likescash.api.product.controller
 import com.dphong.likescash.BaseWebMvcTest
 import com.dphong.likescash.api.product.model.ProductDetails
 import com.dphong.likescash.api.product.service.ProductGetter
-import com.dphong.likescash.common.reponse.DataResult
+import com.dphong.likescash.common.response.DataResult
 import com.dphong.likescash.domain.product.Product
 import com.dphong.likescash.fixtures.MemberFixture
 import org.junit.jupiter.api.Test
