+++ src/test/kotlin/com/dphong/likescash/api/product/controller/RegisterProductControllerTest.kt
@@ -5,7 +5,7 @@ import com.dphong.likescash.api.product.model.RegisterProductRequest
 import com.dphong.likescash.api.product.model.RegisterProductResponse
 import com.dphong.likescash.api.product.model.RegisterProductStatus
 import com.dphong.likescash.api.product.service.RegisterProductService
-import com.dphong.likescash.common.reponse.StatusDataResult
+import com.dphong.likescash.common.response.StatusDataResult
 import com.dphong.likescash.mock.FakeAuthentication
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.junit.jupiter.api.Test
